daytona/requirements.txt
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
numpy==2.4.1
daytona
snowflake-connector-python==3.12.0
python-dotenv==1.0.1
